refactoring builders
