+++ app/src/main/java/com/pedro/rtmpstreamer/customexample/RtspActivity.java
@@ -20,7 +20,7 @@ import android.widget.RadioButton;
 import android.widget.RadioGroup;
 import android.widget.Spinner;
 import android.widget.Toast;
-import com.pedro.builder.RtspBuilder;
+import com.pedro.builder.rtsp.RtspBuilder;
 import com.pedro.encoder.input.video.CameraOpenException;
 import com.pedro.encoder.input.video.EffectManager;
 import com.pedro.rtmpstreamer.R;
