+++ app/src/main/java/com/pedro/rtmpstreamer/defaultexample/ExampleRtspActivity.java
@@ -8,7 +8,7 @@ import android.view.WindowManager;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
-import com.pedro.builder.RtspBuilder;
+import com.pedro.builder.rtsp.RtspBuilder;
 import com.pedro.rtmpstreamer.R;
 import com.pedro.rtsp.rtsp.Protocol;
 import com.pedro.rtsp.utils.ConnectCheckerRtsp;
