+++ app/src/main/java/com/pedro/rtmpstreamer/filestreamexample/RtspFromFileActivity.java
@@ -10,7 +10,7 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
-import com.pedro.builder.RtspBuilderFromFile;
+import com.pedro.builder.rtsp.RtspBuilderFromFile;
 import com.pedro.encoder.input.decoder.VideoDecoderInterface;
 import com.pedro.rtmpstreamer.R;
 import com.pedro.rtsp.rtsp.Protocol;
