+++ app/src/main/java/com/pedro/rtmpstreamer/surfamoceexample/SurfaceModeRtspActivity.java
@@ -10,7 +10,7 @@ import android.view.WindowManager;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
-import com.pedro.builder.RtspBuilderSurfaceMode;
+import com.pedro.builder.rtsp.RtspBuilderSurfaceMode;
 import com.pedro.rtmpstreamer.R;
 import com.pedro.rtsp.rtsp.Protocol;
 import com.pedro.rtsp.utils.ConnectCheckerRtsp;
