--- builder/src/main/java/com/pedro/builder/RtmpBuilder.java
@@ -1,192 +1,0 @@
-package com.pedro.builder;
-
-import android.graphics.ImageFormat;
-import android.hardware.Camera;
-import android.media.MediaCodec;
-import android.os.Build;
-import android.util.Log;
-import android.view.SurfaceView;
-import com.pedro.encoder.audio.AudioEncoder;
-import com.pedro.encoder.audio.GetAccData;
-import com.pedro.encoder.input.audio.GetMicrophoneData;
-import com.pedro.encoder.input.audio.MicrophoneManager;
-import com.pedro.encoder.input.video.Camera1ApiManager;
-import com.pedro.encoder.input.video.CameraOpenException;
-import com.pedro.encoder.input.video.EffectManager;
-import com.pedro.encoder.input.video.GetCameraData;
-import com.pedro.encoder.video.FormatVideoEncoder;
-import com.pedro.encoder.video.GetH264Data;
-import com.pedro.encoder.video.VideoEncoder;
-import java.nio.ByteBuffer;
-import java.util.ArrayList;
-import java.util.List;
-import net.ossrs.rtmp.ConnectCheckerRtmp;
-import net.ossrs.rtmp.SrsFlvMuxer;
-
-/**
- * Created by pedro on 25/01/17.
- */
-
-public class RtmpBuilder implements GetAccData, GetCameraData, GetH264Data, GetMicrophoneData {
-
-  private int width;
-  private int height;
-  private Camera1ApiManager cameraManager;
-  private VideoEncoder videoEncoder;
-  private MicrophoneManager microphoneManager;
-  private AudioEncoder audioEncoder;
-  private SrsFlvMuxer srsFlvMuxer;
-  private boolean streaming;
-  private boolean videoEnabled = true;
-
-  public RtmpBuilder(SurfaceView surfaceView, ConnectCheckerRtmp connectChecker) {
-    cameraManager = new Camera1ApiManager(surfaceView, this);
-    videoEncoder = new VideoEncoder(this);
-    microphoneManager = new MicrophoneManager(this);
-    audioEncoder = new AudioEncoder(this);
-    srsFlvMuxer = new SrsFlvMuxer(connectChecker);
-    streaming = false;
-  }
-
-  public void setAuthorization(String user, String password) {
-    srsFlvMuxer.setAuthorization(user, password);
-  }
-
-  public boolean prepareVideo(int width, int height, int fps, int bitrate, boolean hardwareRotation,
-      int rotation) {
-    this.width = width;
-    this.height = height;
-    int imageFormat = ImageFormat.NV21; //supported nv21 and yv12
-    cameraManager.prepareCamera(width, height, fps, imageFormat);
-    videoEncoder.setImageFormat(imageFormat);
-    return videoEncoder.prepareVideoEncoder(width, height, fps, bitrate, rotation, hardwareRotation,
-        FormatVideoEncoder.YUV420Dynamical);
-  }
-
-  public boolean prepareAudio(int bitrate, int sampleRate, boolean isStereo, boolean echoCanceler,
-      boolean noiseSuppressor) {
-    microphoneManager.createMicrophone(sampleRate, isStereo, echoCanceler, noiseSuppressor);
-    srsFlvMuxer.setIsStereo(isStereo);
-    srsFlvMuxer.setSampleRate(sampleRate);
-    return audioEncoder.prepareAudioEncoder(bitrate, sampleRate, isStereo);
-  }
-
-  public boolean prepareVideo() {
-    cameraManager.prepareCamera();
-    width = videoEncoder.getWidth();
-    height = videoEncoder.getHeight();
-    return videoEncoder.prepareVideoEncoder();
-  }
-
-  public boolean prepareAudio() {
-    microphoneManager.createMicrophone();
-    return audioEncoder.prepareAudioEncoder();
-  }
-
-  public void startStream(String url) {
-    srsFlvMuxer.start(url);
-    srsFlvMuxer.setVideoResolution(width, height);
-    videoEncoder.start();
-    audioEncoder.start();
-    cameraManager.start();
-    microphoneManager.start();
-    streaming = true;
-  }
-
-  public void stopStream() {
-    cameraManager.stop();
-    microphoneManager.stop();
-    srsFlvMuxer.stop();
-    videoEncoder.stop();
-    audioEncoder.stop();
-    streaming = false;
-  }
-
-  public List<String> getResolutions() {
-    List<Camera.Size> list = cameraManager.getPreviewSize();
-    List<String> resolutions = new ArrayList<>();
-    for (Camera.Size size : list) {
-      resolutions.add(size.width + "X" + size.height);
-    }
-    return resolutions;
-  }
-
-  public void disableAudio() {
-    microphoneManager.mute();
-  }
-
-  public void enableAudio() {
-    microphoneManager.unMute();
-  }
-
-  public boolean isAudioMuted() {
-    return microphoneManager.isMuted();
-  }
-
-  public boolean isVideoEnabled() {
-    return videoEnabled;
-  }
-
-  public void disableVideo() {
-    videoEncoder.startSendBlackImage();
-    videoEnabled = false;
-  }
-
-  public void enableVideo() {
-    videoEncoder.stopSendBlackImage();
-    videoEnabled = true;
-  }
-
-  public void switchCamera() throws CameraOpenException {
-    if (isStreaming()) {
-      cameraManager.switchCamera();
-    }
-  }
-
-  /** need min API 19 */
-  public void setVideoBitrateOnFly(int bitrate) {
-    if (Build.VERSION.SDK_INT >= 19) {
-      videoEncoder.setVideoBitrateOnFly(bitrate);
-    }
-  }
-
-  public boolean isStreaming() {
-    return streaming;
-  }
-
-  public void setEffect(EffectManager effect) {
-    if (isStreaming()) {
-      cameraManager.setEffect(effect);
-    }
-  }
-
-  @Override
-  public void getAccData(ByteBuffer accBuffer, MediaCodec.BufferInfo info) {
-    srsFlvMuxer.sendAudio(accBuffer, info);
-  }
-
-  @Override
-  public void onSPSandPPS(ByteBuffer sps, ByteBuffer pps) {
-    srsFlvMuxer.setSpsPPs(sps, pps);
-  }
-
-  @Override
-  public void getH264Data(ByteBuffer h264Buffer, MediaCodec.BufferInfo info) {
-    srsFlvMuxer.sendVideo(h264Buffer, info);
-  }
-
-  @Override
-  public void inputPcmData(byte[] buffer, int size) {
-    audioEncoder.inputPcmData(buffer, size);
-  }
-
-  @Override
-  public void inputYv12Data(byte[] buffer) {
-    videoEncoder.inputYv12Data(buffer);
-  }
-
-  @Override
-  public void inputNv21Data(byte[] buffer) {
-    videoEncoder.inputNv21Data(buffer);
-  }
-}
