+++ builder/src/main/java/com/pedro/builder/base/BuilderBase.java
@@ -1,105 +1,103 @@
-package com.pedro.builder;
+package com.pedro.builder.base;
 
 import android.graphics.ImageFormat;
+import android.hardware.Camera;
 import android.media.MediaCodec;
 import android.os.Build;
-import android.support.annotation.RequiresApi;
 import android.view.SurfaceView;
 import com.pedro.encoder.audio.AudioEncoder;
 import com.pedro.encoder.audio.GetAccData;
 import com.pedro.encoder.input.audio.GetMicrophoneData;
 import com.pedro.encoder.input.audio.MicrophoneManager;
-import com.pedro.encoder.input.video.Camera2ApiManager;
+import com.pedro.encoder.input.video.Camera1ApiManager;
 import com.pedro.encoder.input.video.CameraOpenException;
+import com.pedro.encoder.input.video.EffectManager;
 import com.pedro.encoder.input.video.GetCameraData;
 import com.pedro.encoder.video.FormatVideoEncoder;
 import com.pedro.encoder.video.GetH264Data;
 import com.pedro.encoder.video.VideoEncoder;
-import com.pedro.rtsp.rtsp.Protocol;
-import com.pedro.rtsp.rtsp.RtspClient;
-import com.pedro.rtsp.utils.ConnectCheckerRtsp;
 import java.nio.ByteBuffer;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
- * Created by pedro on 4/06/17.
- * This builder is under test, rotation only work with hardware because use encoding surface mode.
+ * Created by pedro on 7/07/17.
  */
-@RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-public class RtspBuilderSurfaceMode implements GetAccData, GetCameraData, GetH264Data, GetMicrophoneData {
 
-  private Camera2ApiManager cameraManager;
+public abstract class BuilderBase
+    implements GetAccData, GetCameraData, GetH264Data, GetMicrophoneData {
+
+  private Camera1ApiManager cameraManager;
   private VideoEncoder videoEncoder;
-  private MicrophoneManager microphoneManager;
-  private AudioEncoder audioEncoder;
+  protected MicrophoneManager microphoneManager;
+  protected AudioEncoder audioEncoder;
   private boolean streaming;
-  private SurfaceView surfaceView;
-
-  private RtspClient rtspClient;
   private boolean videoEnabled = true;
 
-  public RtspBuilderSurfaceMode(SurfaceView surfaceView, Protocol protocol,
-      ConnectCheckerRtsp connectCheckerRtsp) {
-    this.surfaceView = surfaceView;
-    rtspClient = new RtspClient(connectCheckerRtsp, protocol);
+  public BuilderBase(SurfaceView surfaceView) {
+    cameraManager = new Camera1ApiManager(surfaceView, this);
     videoEncoder = new VideoEncoder(this);
     microphoneManager = new MicrophoneManager(this);
     audioEncoder = new AudioEncoder(this);
     streaming = false;
   }
 
-  public void setAuthorization(String user, String password) {
-    rtspClient.setAuthorization(user, password);
+  public abstract void setAuthorization(String user, String password);
+
+  public boolean prepareVideo(int width, int height, int fps, int bitrate, boolean hardwareRotation,
+      int rotation) {
+    int imageFormat = ImageFormat.NV21; //supported nv21 and yv12
+    cameraManager.prepareCamera(width, height, fps, imageFormat);
+    videoEncoder.setImageFormat(imageFormat);
+    return videoEncoder.prepareVideoEncoder(width, height, fps, bitrate, rotation, hardwareRotation,
+        FormatVideoEncoder.YUV420Dynamical);
   }
 
-  public boolean prepareVideo(int width, int height, int fps, int bitrate, int rotation) {
-    int imageFormat = ImageFormat.NV21; //supported nv21 and yv12
-    videoEncoder.setImageFormat(imageFormat);
-    boolean result = videoEncoder.prepareVideoEncoder(width, height, fps, bitrate, rotation,
-        true, FormatVideoEncoder.SURFACE);
-    cameraManager = new Camera2ApiManager(surfaceView, videoEncoder.getInputSurface(), surfaceView.getContext());
-    return result;
-  }
+  protected abstract void prepareAudioRtp(boolean isStereo, int sampleRate);
 
   public boolean prepareAudio(int bitrate, int sampleRate, boolean isStereo, boolean echoCanceler,
       boolean noiseSuppressor) {
-    rtspClient.setSampleRate(sampleRate);
-    rtspClient.setIsStereo(isStereo);
     microphoneManager.createMicrophone(sampleRate, isStereo, echoCanceler, noiseSuppressor);
+    prepareAudioRtp(isStereo, sampleRate);
     return audioEncoder.prepareAudioEncoder(bitrate, sampleRate, isStereo);
   }
 
   public boolean prepareVideo() {
-    boolean result = videoEncoder.prepareVideoEncoder(640, 480, 30, 1200 * 1024, 90,
-        true, FormatVideoEncoder.SURFACE);
-    cameraManager = new Camera2ApiManager(surfaceView, videoEncoder.getInputSurface(), surfaceView.getContext());
-    return result;
+    cameraManager.prepareCamera();
+    return videoEncoder.prepareVideoEncoder();
   }
 
-  //set 16000hz sample rate because 44100 produce desynchronization audio in rtsp
-  public boolean prepareAudio() {
-    microphoneManager.setSampleRate(16000);
-    audioEncoder.setSampleRate(16000);
-    microphoneManager.createMicrophone();
-    rtspClient.setSampleRate(microphoneManager.getSampleRate());
-    return audioEncoder.prepareAudioEncoder();
-  }
+  public abstract boolean prepareAudio();
+
+  protected abstract void startStreamRtp(String url);
 
   public void startStream(String url) {
+    startStreamRtp(url);
     videoEncoder.start();
     audioEncoder.start();
-    cameraManager.openCameraBack();
+    cameraManager.start();
     microphoneManager.start();
-    rtspClient.setUrl(url);
     streaming = true;
   }
 
+  protected abstract void stopStreamRtp();
+
   public void stopStream() {
-    rtspClient.disconnect();
-    cameraManager.closeCamera();
+    cameraManager.stop();
     microphoneManager.stop();
+    stopStreamRtp();
     videoEncoder.stop();
     audioEncoder.stop();
     streaming = false;
+  }
+
+  public List<String> getResolutions() {
+    List<Camera.Size> list = cameraManager.getPreviewSize();
+    List<String> resolutions = new ArrayList<>();
+    for (Camera.Size size : list) {
+      resolutions.add(size.width + "X" + size.height);
+    }
+    return resolutions;
   }
 
   public void disableAudio() {
@@ -110,6 +108,14 @@ public class RtspBuilderSurfaceMode implements GetAccData, GetCameraData, GetH26
     microphoneManager.unMute();
   }
 
+  public boolean isAudioMuted() {
+    return microphoneManager.isMuted();
+  }
+
+  public boolean isVideoEnabled() {
+    return videoEnabled;
+  }
+
   public void disableVideo() {
     videoEncoder.startSendBlackImage();
     videoEnabled = false;
@@ -118,14 +124,6 @@ public class RtspBuilderSurfaceMode implements GetAccData, GetCameraData, GetH26
   public void enableVideo() {
     videoEncoder.stopSendBlackImage();
     videoEnabled = true;
-  }
-
-  public boolean isAudioMuted() {
-    return microphoneManager.isMuted();
-  }
-
-  public boolean isVideoEnabled() {
-    return videoEnabled;
   }
 
   public void switchCamera() throws CameraOpenException {
@@ -145,32 +143,31 @@ public class RtspBuilderSurfaceMode implements GetAccData, GetCameraData, GetH26
     return streaming;
   }
 
-  //public void setEffect(EffectManager effect) {
-  //  if (isStreaming()) {
-  //    cameraManager.setEffect(effect);
-  //  }
-  //}
+  public void setEffect(EffectManager effect) {
+    if (isStreaming()) {
+      cameraManager.setEffect(effect);
+    }
+  }
+
+  protected abstract void getAccDataRtp(ByteBuffer accBuffer, MediaCodec.BufferInfo info);
 
   @Override
   public void getAccData(ByteBuffer accBuffer, MediaCodec.BufferInfo info) {
-    rtspClient.sendAudio(accBuffer, info);
+    getAccDataRtp(accBuffer, info);
   }
+
+  protected abstract void onSPSandPPSRtp(ByteBuffer sps, ByteBuffer pps);
 
   @Override
   public void onSPSandPPS(ByteBuffer sps, ByteBuffer pps) {
-    byte[] mSPS = new byte[sps.capacity() - 4];
-    sps.position(4);
-    sps.get(mSPS, 0, mSPS.length);
-    byte[] mPPS = new byte[pps.capacity() - 4];
-    pps.position(4);
-    pps.get(mPPS, 0, mPPS.length);
-    rtspClient.setSPSandPPS(mPPS, mSPS);
-    rtspClient.connect();
+    onSPSandPPSRtp(sps, pps);
   }
+
+  protected abstract void getH264DataRtp(ByteBuffer h264Buffer, MediaCodec.BufferInfo info);
 
   @Override
   public void getH264Data(ByteBuffer h264Buffer, MediaCodec.BufferInfo info) {
-    rtspClient.sendVideo(h264Buffer, info);
+    getH264DataRtp(h264Buffer, info);
   }
 
   @Override
@@ -187,6 +184,4 @@ public class RtspBuilderSurfaceMode implements GetAccData, GetCameraData, GetH26
   public void inputNv21Data(byte[] buffer) {
     videoEncoder.inputNv21Data(buffer);
   }
-
 }
-
