+++ builder/src/main/java/com/pedro/builder/rtmp/RtmpBuilderFromFile.java
@@ -1,0 +1,58 @@
+package com.pedro.builder.rtmp;
+
+/**
+ * Created by pedro on 26/06/17.
+ */
+
+import android.media.MediaCodec;
+import android.os.Build;
+import android.support.annotation.RequiresApi;
+import com.pedro.builder.base.BuilderFromFileBase;
+import com.pedro.encoder.input.decoder.VideoDecoderInterface;
+import java.nio.ByteBuffer;
+import net.ossrs.rtmp.ConnectCheckerRtmp;
+import net.ossrs.rtmp.SrsFlvMuxer;
+
+/**
+ * Created by pedro on 26/06/17.
+ * This builder is under test, rotation only work with hardware because use encoding surface mode.
+ * Only video is working, audio will be added when it work
+ */
+@RequiresApi(api = Build.VERSION_CODES.JELLY_BEAN_MR2)
+public class RtmpBuilderFromFile extends BuilderFromFileBase {
+
+  private SrsFlvMuxer srsFlvMuxer;
+
+  public RtmpBuilderFromFile(ConnectCheckerRtmp connectChecker,
+      VideoDecoderInterface videoDecoderInterface) {
+    super(videoDecoderInterface);
+    srsFlvMuxer = new SrsFlvMuxer(connectChecker);
+  }
+
+  @Override
+  public void setAuthorization(String user, String password) {
+    srsFlvMuxer.setAuthorization(user, password);
+  }
+
+  @Override
+  protected void startStreamRtp(String url) {
+    srsFlvMuxer.start(url);
+  }
+
+  @Override
+  protected void stopStreamRtp() {
+    srsFlvMuxer.stop();
+  }
+
+  @Override
+  protected void onSPSandPPSRtp(ByteBuffer sps, ByteBuffer pps) {
+    srsFlvMuxer.setSpsPPs(sps, pps);
+  }
+
+  @Override
+  protected void getH264DataRtp(ByteBuffer h264Buffer, MediaCodec.BufferInfo info) {
+    srsFlvMuxer.sendVideo(h264Buffer, info);
+  }
+}
+
+
